Switch package name for class+role composed class to be valid
We used to join together the components with + so you got -

  Class+Role1+Role2+Role3

however this then screws with Sub::Quote's ability to generate a constructor
into the final class since

  sub Class+Role1+Role2+Role3::new {

isn't valid perl. Therefore, I've switched the code to generate

  Class__WITH__Role1__AND__Role3__AND__Role3

which is substantially uglier but works. Thanks to DGL for the test