Tenative switch to a generated DEMOLISHALL - see rest of message for caveats
I'm wondering if the reason that ajgb didn't do it this way is that it has
bad interactions with global destruction on some perls; it did, however, work
fine in the t/ cases. Rather than use the ugly "let's figure this out every
time" code that he wrote (which did, admittedly, work) I think that if the
generation doesn't work across perls then our best path forwards is to try
and accelerate DEMOLISHALL construction - perhaps even lifting it to new()
time so that classes without a DEMOLISH method don't get a DESTROY added (I
hate having to always have a DESTROY method).

But first, I want to see if this breaks so that I know how to test that
lifting to new() time fixes it.