Adds moosify isa spec
This feature is an optional coderef that gets fired when Moo::HandleMoose is
injecting a real metaclass. The spec hashref is passed as the only argument to
allow for modifications to carry forward to Moose.

The main usecase for this is to allow thirdparty extensions such as
MooX::HandlesVia to mutate their Moo specific attributes to something Moose
will understand.