fix ->does on Moose child class that composes roles
Moose will not override Moo's does method, so calling ->does on a child
class will not check against that child's meta information.  Moo doesn't
know about the roles composed into the child class, so it will return
false when it should return true.

When Moo's does is called on a Moose class, check the Moose metaclass
for role information.