Fix inflating MGC by moving 'use Moo;' to top of file (RT#101111)
Moo tracks all existing subs as 'not methods' when it is imported.
These methods need to be tracked properly for Moose metaclass inflation
to work, which can happen for any Moo class.  Although
Method::Generate::Constructor doesn't have any accessors installed, we
do list a number of attributes to allow subclassing with Moo to work.
The accessors for those attributes have to be considered methods.  We can
'use Moo;' at the top of MGC without causing issues with bootstrapping
because a constructor won't be installed by the import line alone.