fix calling methods on FakeMetaClass class
FakeMetaClass normally proxyies to a real metaclass that it instantiates
if needed.  That only works for object instances though.  For methods
called on the class, the only sensible thing to do is replicate how it
would act without the proxying.

This can be important if a module scans all packages, calling can on all
of them.  An example of this if forks.pm, scanning for CLONE_SKIP.