check for existence of sub naming subs before using them
Sub::Util ships with new versions of List::Util, but its subs are
provided by the XS component of List::Util.  All Sub::Name itself does
is load List::Util to ensure the subs are available.

If List::Util is downgraded or the new version is otherwise unavailable,
Sub::Util will load fine but the subs won't be installed.  We need to
make sure the actual subs exist before we try to use them.