testcase for init_arg with spaces in it
Currently Moo fails to construct an object if there is
an attribute with an init_arg with spaces in it which is
also 'required'.

While keys with spaces are not a sensible thing, sometimes
it is nice to be able to construct objects from hashes
built by existing code that have wacky keys.