allow is => lazy, required => 1, init_arg => undef
We should be asking the accessor generator if an attribute has a
default, not checking in the constructor generator.  The accessor
generator knows about the various ways to specify a default.

It would eventually be preferrable to have the is => 'lazy' to builder
=> "_build_$attr" transformation done in a reusable place, so we don't
have to continue to check for the shortcut.