meta inflation shouldn't install deferred subs after undeferring
We want to have stable subs so that the meta model is accurate, so we
undefer all subs before populating the meta object.  But we need to be
sure to use the undeferred subs when populating, or we'll reinstall the
deferred subs.  That would mean all of the method calls having to pay
for a goto, as well as just being strange and confusing.

This was also covering up a bug where the constructor generator was
intending to not allow replacing an inlined constructor, but allowed
replacing this "half inlined" form.