make all $] checks consistent
In some versions of perl, the floating point number in $] is not
consistent with how a numeric literal will be stored, so it won't
compare correctly with numeric comparisons.  Stringify $] first, so it
will be compared correctly.  For consistency, do this even for cases
where it isn't required, as well as using full length versions split
with underscores.