avoid passing slash in first parameter in makefile
If the command is in quotes and the first parameter has a slash, nmake
adds a space before the slash (!?!).  Fix this by reversing the order we
use the parameters.  -I additions to @INC take effect before -M modules
are loaded, so reversing them is fine.

See also commit 006d552 in ExtUtils-MakeMaker.