fix Type::Tiny inflation on old threaded perl
%TYPE_MAP defines how to inflate types from Moo types (code refs) to
Moose types.  The keys are the code refs, which means they get
stringified.

When using threads, the refaddr of refs will change.  This means the
stringification will change.  Because of this, we tie %TYPE_MAP so we
can keep real references to the types and fix their stringification when
cloning into a new thread.  However, %TYPE_MAP is a global that can be
added to without loading Moo.  If Moo is loaded later, it needs to do
something with the existing stringified refs in the mapping.  It handles
this by using an ugly B hack to turn the refaddr into the real
reference.

perl versions before 5.10 attach overloading magic to the reference
rather than the referent.  When restoring the reference using B, it
doesn't re-attach the overloading magic.

Type::Tiny defines a custom stringification which mimics a standard perl
ref stringification, but it uses slightly different formatting.  The
refaddr will represent the same integer, but won't always have the same
length.  Because it looks like a normal ref, Moo performs its string to
ref inflation using B.  But because B doesn't add the overloading magic
to the ref used in inflation, some references to the type will use
Type::Tiny's custom stringification, and some will use perl's standard
stringification.  This will prevent some operations from matching the
strings, and the custom inflation won't trigger.

Reblessing the reference after retrieving it from B will properly apply
the overloading magic, fixing the issue.