fix error location test on perls with Carp::Heavy
Our tracing code involves clearing %Carp::Internal and
%Carp::CarpInternal temporarily with local. If our call to longmess is
then the first thing to call Carp, it will load Carp::Heavy on versions
that use it. Carp::Heavy is what initially sets up %Carp::Internal and
%Carp::CarpInternal, so their initialization will be lost when our local
is undone. This would mean Carp itself was not listed in the variables,
so Carp wouldn't be able to track the first caller properly, and croak
would essentially work the same as die. This was the intended behavior
of our longmess call in the tracing code, but it wasn't meant to impact
other calls.

Load Carp::Heavy early to ensure it is properly intialized, so our
localization will restore the properly initialized variables.