drop MRO::Compat prereq, using our own isa calculations
We don't care about changing how methods are resolved, only getting the
linearized isa tree. The user would need to either use mro or
MRO::Compat to set an alternate MRO. If the user has set an alternate
MRO, this means mro::get_linear_isa will be available, and we can use
it. If it isn't available, the MRO will have to be DFS, which we can
easily implement ourselves.

This allows us to eliminate the conditional dependency on MRO::Compat,
and only needing a small amount of back compat code.