revise global destruction checks to remove mandatory external dependencies
On newer perls, we use ${^GLOBAL_PHASE} to detect global destruction.
Older perls required the Devel::GlobalDestruction module to check this.
However, the code it uses to check this is tiny and we can easily inline
it. We still use Devel::GlobalDestruction::XS if available for
performance.

When fatpacking code that uses Moo on a newer perl,
Devel::GlobalDestruction won't be used and may not be installed, unless
the user adds it themselves. This means the fatpacked code won't work on
older perls. Removing our hard dependency on the module means fatpacked
code from any perl version should work on older perl versions.

We still list Devel::GlobalDestruction as a dependency on older perls.
This will bring in Devel::GlobalDestruction::XS when possible, so we
don't have to do a compiler check ourselves.