fix populating %INC on create_class_with_roles
create_class_with_roles generates a class and then sets an appropriate
entry in %INC for it. Previously, it would set the value to the file it
was called from. After refactoring, the direct caller is Role::Tiny, not
the user code. We need to search up the call stack for the appropriate
file to set.