force coerce builder to a coderef when installing
We want to preserve the original arguments passed to the accessor
generator where possible. We check various arguments to ensure that they
can be converted to a coderef. But when installing a builder sub, it has
to be the coderef, not the object, that gets installed into the symbol
table.